requirements.txt
asgiref==3.8.1
Django==5.0.3
sqlparse==0.4.4
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.1
